fix: rename logging.py to log_config.py to avoid shadowing stdlib logging module
