+++ app/main.py
@@ -13,7 +13,7 @@ from app.agent.prompts import SYSTEM_PROMPT
 from app.agent.tools import BDToolkit
 from app.api.main import api_router
 from app.db.database import engine, init_database
-from app.logging import setup_logger
+from app.log_config import setup_logger
 from app.settings import settings
 
 setup_logger()
